@@ -23,26 +23,32 @@ environment:
 dependencies:
   flutter:
     sdk: flutter
-  http: ^0.13.3
-  provider: ^5.0.0
-  shared_preferences: ^2.0.6
+  http: ^0.13.4
+  provider: ^6.0.1
+  shared_preferences: ^2.0.8
   uri: ^1.0.0
-  webview_flutter: ^2.0.10
+  webview_flutter: ^2.1.2
   html: ^0.15.0
-  url_launcher: ^6.0.9
-  json_annotation: ^4.1.0
+  url_launcher: ^6.0.12
+  json_annotation: ^4.3.0
+  cached_network_image: ^3.1.0
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.3
 
 dev_dependencies:
+  integration_test:
+    sdk: flutter
   flutter_test:
     sdk: flutter
-  build_runner: ^2.0.6
+  build_runner: ^2.1.4
   network_image_mock: ^2.0.1
-  mocktail: ^0.1.4
-  json_serializable: ^5.0.0
+  mocktail: ^0.2.0
+  json_serializable: ^6.0.1
+  flutter_dotenv: ^5.0.2
+  flutter_lints: ^1.0.4
+  flutter_gherkin: ^3.0.0-rc.6
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
@@ -56,7 +62,8 @@ flutter:
   uses-material-design: true
 
   # To add assets to your application, add an assets section, like this:
-  # assets:
+  assets:
+    - .env
   #   - images/a_dot_burr.jpeg
   #   - images/a_dot_ham.jpeg
 
