@@ -1,15 +1,32 @@
 import 'package:brick_app/model/brick_set_list.dart';
 import 'package:brick_app/model/rebrickable_model.dart';
 import 'package:brick_app/pages/set_list_page.dart';
+import 'package:brick_app/pages/utils.dart';
 import 'package:brick_app/widgets/brick_app_bar.dart';
-import 'package:flutter/cupertino.dart';
+import 'package:brick_app/widgets/create_delete_dialog.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 
-class OverviewPage extends StatelessWidget {
+class OverviewPage extends StatefulWidget {
+  const OverviewPage({Key? key}) : super(key: key);
+
+  @override
+  State<StatefulWidget> createState() => _OverviewPageState();
+}
+
+class _OverviewPageState extends State<OverviewPage> {
+  Future<List<BrickSetList>>? _brickSetListsFuture;
+
+  @override
+  initState() {
+    super.initState();
+    _refreshBrickSetList(context);
+  }
+
+  @override
   Widget build(BuildContext context) {
     return FutureBuilder<List<BrickSetList>>(
-        future: context.read<RebrickableModel>().getUsersSetLists(),
+        future: _brickSetListsFuture,
         builder: (context, snapshot) {
           if (snapshot.hasData) {
             return Scaffold(
@@ -19,22 +36,40 @@ class OverviewPage extends StatelessWidget {
               body: Center(
                 child: _createListView(snapshot.data!),
               ),
+              floatingActionButton: FloatingActionButton(
+                  key: const Key('createSetList'),
+                  child: const Icon(Icons.add),
+                  onPressed: () => {
+                        showInputDialog(context,
+                            title: 'Create new Set List',
+                            inputFieldKey: const Key('setListName'),
+                            label: 'Set list name',
+                            okButtonText: 'Create',
+                            onOkButtonPress: (input) async {
+                          await context
+                              .read<RebrickableModel>()
+                              .addSetList(setListName: input);
+                          await _refreshBrickSetList(context);
+                          Navigator.of(context).pop();
+                          showSnackBar(context, 'List created successfully');
+                        }),
+                      }),
             );
           } else if (snapshot.hasError) {
             return Scaffold(
               appBar: AppBar(
-                title: Text('My Set Lists (error)'),
+                title: const Text('My Set Lists (error)'),
               ),
-              body: Center(
+              body: const Center(
                 child: Text('An error occured while loading the set lists.'),
               ),
             );
           } else {
             return Scaffold(
               appBar: AppBar(
-                title: Text('My Set Lists (loading)'),
+                title: const Text('My Set Lists (loading)'),
               ),
-              body: Center(
+              body: const Center(
                 child: CircularProgressIndicator(),
               ),
             );
@@ -42,25 +77,61 @@ class OverviewPage extends StatelessWidget {
         });
   }
 
-  ListView _createListView(List<BrickSetList> brickSetLists) =>
-      brickSetLists.length != 0
+  Widget _createListView(List<BrickSetList> brickSetLists) => RefreshIndicator(
+      child: brickSetLists.isNotEmpty
           ? ListView.builder(
-              key: ObjectKey('setList'),
+              key: const Key('overviewList'),
               itemBuilder: (context, index) =>
                   _createListTile(context, brickSetLists[index]),
               itemCount: brickSetLists.length)
           : ListView(
-              children: [Text('You have no set lists in your account.')]);
+              children: const [
+                Text('You have no set lists in your account.'),
+              ],
+            ),
+      onRefresh: () => _refreshBrickSetList(context));
 
   ListTile _createListTile(BuildContext context, BrickSetList brickSetList) =>
       ListTile(
-        leading: Icon(
+        key: Key('overviewListTile_${brickSetList.name}'),
+        leading: const Icon(
           Icons.domain,
           color: Colors.red,
         ),
         title: Text(brickSetList.name),
         subtitle: Text('${brickSetList.numSets} sets'),
-        onTap: () => Navigator.of(context).push(MaterialPageRoute(
-            builder: (context) => SetListPage(brickSetList: brickSetList))),
+        onTap: () => Navigator.of(context).push(
+          MaterialPageRoute(
+            builder: (context) => SetListPage(brickSetList: brickSetList),
+          ),
+        ),
+        trailing: IconButton(
+          key: Key('deleteSetList_${brickSetList.name}'),
+          icon: const Icon(Icons.delete),
+          onPressed: () => showDialog(
+            context: context,
+            builder: (context) => CreateDeleteDialog(
+              title: 'Delete Set List',
+              content:
+                  const Text('''Do you really want to delete this set list? 
+This deletes the list itself and all sets in this list.'''),
+              okButtonText: 'Delete',
+              onOkButtonPress: () async {
+                final model = context.read<RebrickableModel>();
+                await model.deleteSetList(setListId: brickSetList.id);
+                await _refreshBrickSetList(context);
+                Navigator.of(context).pop();
+                showSnackBar(context, 'List deleted successfully');
+              },
+            ),
+          ),
+        ),
       );
+
+  Future<void> _refreshBrickSetList(BuildContext context) async {
+    setState(() {
+      _brickSetListsFuture =
+          context.read<RebrickableModel>().getUsersSetLists();
+    });
+  }
 }
