@@ -53,7 +53,7 @@ class Dummy {
   Dummy(this.name);
 }
 
-final firstUrlResponseBody = '''{
+const firstUrlResponseBody = '''{
   "count": 39932,
   "next": "https://myUrl?page=2",
   "previous": null,
@@ -62,7 +62,7 @@ final firstUrlResponseBody = '''{
   ]
 }''';
 
-final secondUrlResponseBody = '''{
+const secondUrlResponseBody = '''{
   "count": 39932,
   "next": "https://myUrl?page=3",
   "previous": "https://myUrl",
@@ -71,7 +71,7 @@ final secondUrlResponseBody = '''{
   ]
 }''';
 
-final thirdUrlResponseBody = '''{
+const thirdUrlResponseBody = '''{
   "count": 39932,
   "next": null,
   "previous": "https://myUrl?page=2",
