@@ -11,13 +11,13 @@ import 'package:mocktail/mocktail.dart';
 import '../mocks.dart';
 
 void main() {
-  final String _apiKey = 'apiKey';
+  const String _apiKey = 'apiKey';
   final Map<String, String> _authHeader = {'Authorization': 'key $_apiKey'};
   final Map<String, String> _authHeaderWithContentType = {
     'Authorization': 'key $_apiKey',
     'content-type': 'application/x-www-form-urlencoded'
   };
-  var client;
+  late Client client;
   late RebrickableService service;
 
   void _setUpAuthenticatedServiceMock() async {
@@ -80,7 +80,7 @@ void main() {
     });
 
     test('it should retrieve a specific set list', () async {
-      final id = 521857;
+      const id = 521857;
       final uri = Uri.parse(userSetListUrlTemplate
           .expand({'user_token': 'validtoken', 'list_id': id}));
       when(() => client.get(uri, headers: _authHeader))
@@ -90,13 +90,79 @@ void main() {
     });
   });
 
+  group('add set list', () {
+    setUp(() async {
+      _setUpAuthenticatedServiceMock();
+    });
+
+    test('executes fine', () async {
+      const setListName = 'setListName';
+      final url =
+          Uri.parse(addSetListUrlTemplate.expand({'user_token': 'validtoken'}));
+      const body = 'is_buildable=true&name=$setListName&num_sets=0';
+      final headers = _authHeaderWithContentType;
+      when(() => client.post(url, headers: headers, body: body))
+          .thenAnswer((_) async => Response('', 201));
+
+      await service.addSetList(setListName: setListName);
+      verify(() => client.post(url, headers: headers, body: body)).called(1);
+    });
+
+    test('handles error', () async {
+      const setListName = 'setListName';
+      final url =
+          Uri.parse(addSetListUrlTemplate.expand({'user_token': 'validtoken'}));
+      const body = 'is_buildable=true&name=$setListName&num_sets=0';
+      final headers = _authHeaderWithContentType;
+      when(() => client.post(url, headers: headers, body: body))
+          .thenAnswer((_) async => Response('', 500));
+
+      expect(() async => await service.addSetList(setListName: setListName),
+          throwsA(isA<RebrickableApiException>()));
+      verify(() => client.post(url, headers: headers, body: body)).called(1);
+    });
+  });
+
+  group('delete set list', () {
+    setUp(() async {
+      _setUpAuthenticatedServiceMock();
+    });
+
+    test('executes fine', () async {
+      const setListId = 1;
+      final url = Uri.parse(deleteSetListUrlTemplate
+          .expand({'user_token': 'validtoken', 'list_id': setListId}));
+      final headers = _authHeader;
+      when(() => client.delete(url, headers: headers))
+          .thenAnswer((_) async => Response('', 204));
+
+      await service.deleteSetList(setListId: setListId);
+
+      verify(() => client.delete(url, headers: headers)).called(1);
+    });
+
+    test('handles error', () async {
+      const setListId = 1;
+      final url = Uri.parse(deleteSetListUrlTemplate
+          .expand({'user_token': 'validtoken', 'list_id': setListId}));
+      final headers = _authHeader;
+      when(() => client.delete(url, headers: headers))
+          .thenAnswer((_) async => Response('', 404));
+
+      expect(() async => await service.deleteSetList(setListId: setListId),
+          throwsA(isA<RebrickableApiException>()));
+
+      verify(() => client.delete(url, headers: headers)).called(1);
+    });
+  });
+
   group('set list details', () {
     setUp(() async {
       _setUpAuthenticatedServiceMock();
     });
 
     test('it should retrieve a specific set', () async {
-      final id = 548040;
+      const id = 548040;
       final uri = Uri.parse(userSetListDetailsUrlTemplate
           .expand({'user_token': 'validtoken', 'list_id': id}));
       when(() => client.get(uri, headers: _authHeader))
@@ -109,7 +175,7 @@ void main() {
     });
 
     test('it should handle http errors', () async {
-      final id = 548040;
+      const id = 548040;
       final uri = Uri.parse(userSetListDetailsUrlTemplate
           .expand({'user_token': 'validtoken', 'list_id': id}));
       when(() => client.get(uri, headers: _authHeader))
@@ -126,7 +192,7 @@ void main() {
     });
 
     test('it should retrieve mocs for a set', () async {
-      final setNum = "70672-1";
+      const setNum = "70672-1";
       final uri = Uri.parse(setMocListUrlTemplate.expand({'set_num': setNum}));
       when(() => client.get(uri, headers: _authHeader))
           .thenAnswer((_) async => Response(aMocList, 200));
@@ -137,7 +203,7 @@ void main() {
     });
 
     test('it should handle http errors', () async {
-      final setNum = "70672-1";
+      const setNum = "70672-1";
       final uri = Uri.parse(setMocListUrlTemplate.expand({'set_num': setNum}));
       when(() => client.get(uri, headers: _authHeader))
           .thenAnswer((_) async => Response('not found', 404));
@@ -153,7 +219,7 @@ void main() {
     });
 
     test('it should retrieve inventories for a set', () async {
-      final setNum = "70672-1";
+      const setNum = "70672-1";
       final uri = Uri.parse(setPartListUrlTemplate.expand({'set_num': setNum}));
       when(() => client.get(uri, headers: _authHeader))
           .thenAnswer((_) async => Response(anInventoryList, 200));
@@ -165,7 +231,7 @@ void main() {
     });
 
     test('it should handle http errors', () async {
-      final setNum = "70672-1";
+      const setNum = "70672-1";
       final uri = Uri.parse(setPartListUrlTemplate.expand({'set_num': setNum}));
       when(() => client.get(uri, headers: _authHeader))
           .thenAnswer((_) async => Response('not found', 404));
@@ -198,7 +264,7 @@ void main() {
   });
 }
 
-final aSetListList = '''{
+const aSetListList = '''{
         "count": 1,
         "next": null,
         "previous": null,
@@ -212,7 +278,7 @@ final aSetListList = '''{
         ]
       }''';
 
-final aSingleSetListList = '''{
+const aSingleSetListList = '''{
           "count": 1,
           "next": null,
           "previous": null,
@@ -226,7 +292,7 @@ final aSingleSetListList = '''{
           ]
         }''';
 
-final aSingleSet = '''{
+const aSingleSet = '''{
           "count": 1,
           "next": null,
           "previous": null,
@@ -249,7 +315,7 @@ final aSingleSet = '''{
           ]
         }''';
 
-final aMocList = '''{
+const aMocList = '''{
           "count": 2,
           "next": null,
           "previous": null,
@@ -279,7 +345,7 @@ final aMocList = '''{
           ]
         }''';
 
-final anInventoryList = '''{
+const anInventoryList = '''{
             "count": 2,
             "next": null,
             "previous": null,
